requirements.txt
gunicorn
django-heroku
django
django-environ
django_crispy_forms
crispy-bootstrap5
django-storages
boto3
pillow
django-localflavor
django-phonenumber-field
phonenumbers
easy-thumbnails 
pandas
channels
django-private-chat2
django-cors-headers
plotly==5.7.0
whitenoise
